Fixes license strings for reference model
