+++ models/experimental/detr3d/reference/helpers.py
@@ -1,14 +1,12 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
-
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 # SPDX-License-Identifier: Apache-2.0
+# Modified from the 3Detr implementation in facebookresearch
+# https://github.com/facebookresearch/3detr
+# Copyright (c) Facebook, Inc. and its affiliates.
 
-import copy
 import torch.nn as nn
 from functools import partial
-
-"""
-Copy-paste from https://github.com/facebookresearch/3detr
-"""
+import copy
 
 
 class BatchNormDim1Swap(nn.BatchNorm1d):
