+++ models/experimental/detr3d/reference/model_config.py
@@ -1,5 +1,4 @@
 # SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
-
 # SPDX-License-Identifier: Apache-2.0
