+++ models/experimental/detr3d/reference/utils/dist.py
@@ -1,4 +1,9 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+# SPDX-License-Identifier: Apache-2.0
+# Modified from the 3Detr implementation in facebookresearch
+# https://github.com/facebookresearch/3detr
 # Copyright (c) Facebook, Inc. and its affiliates.
+
 import pickle
 
 import torch
