+++ models/experimental/detr3d/reference/utils/pc_util.py
@@ -1,23 +1,17 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+# SPDX-License-Identifier: Apache-2.0
+# Modified from the 3Detr implementation in facebookresearch
+# https://github.com/facebookresearch/3detr
 # Copyright (c) Facebook, Inc. and its affiliates.
 
-""" Utility functions for processing point clouds.
-
-Author: Charles R. Qi and Or Litany
-"""
-
 import torch
-
-# Point cloud IO
+import trimesh
 import numpy as np
 
-# Mesh IO
-import trimesh
 
 # ----------------------------------------
 # Point Cloud Sampling
 # ----------------------------------------
-
-
 def random_sampling(pc, num_sample, replace=None, return_choices=False):
     """Input is NxC, output is num_samplexC"""
     if replace is None:
