+++ models/experimental/detr3d/ttnn/model_3detr.py
@@ -8,7 +8,7 @@ import torch
 import numpy as np
 
 from models.common.lightweightmodule import LightweightModule
-from models.experimental.detr3d.reference.model_utils import BoxProcessor
+from models.experimental.detr3d.reference.model_3detr import BoxProcessor
 from models.experimental.detr3d.ttnn.masked_transformer_encoder import (
     TtnnTransformerEncoderLayer,
     TtnnMaskedTransformerEncoder,
@@ -21,7 +21,7 @@ from models.experimental.detr3d.ttnn.transformer_decoder import (
 )
 from models.experimental.detr3d.ttnn.generic_mlp import TtnnGenericMLP
 from models.experimental.detr3d.ttnn.pointnet_samodule_votes import TtnnPointnetSAModuleVotes
-from models.experimental.detr3d.reference.torch_pointnet2_ops import FurthestPointSampling
+from models.experimental.detr3d.reference.torch_pointnet2_ops import furthest_point_sample
 from models.experimental.detr3d.ttnn.position_embedding import TtnnPositionEmbeddingCoordsSine
 
 
@@ -82,7 +82,6 @@ class TtnnModel3DETR(LightweightModule):
         self.num_queries = num_queries
         self.torch_box_processor = BoxProcessor(dataset_config)
         # self.box_processor = TtnnBoxProcessor(dataset_config, device=self.device)
-        self.torch_furthest_point_sample = FurthestPointSampling()
 
     def build_mlp_heads(self):
         self.mlp_heads = {
@@ -109,7 +108,7 @@ class TtnnModel3DETR(LightweightModule):
         }
 
     def get_query_embeddings(self, torch_encoder_xyz, point_cloud_dims):
-        torch_query_inds = self.torch_furthest_point_sample(torch_encoder_xyz, self.num_queries)
+        torch_query_inds = furthest_point_sample(torch_encoder_xyz, self.num_queries)
         torch_query_inds = torch_query_inds.long()
         torch_query_xyz = [torch.gather(torch_encoder_xyz[..., x], 1, torch_query_inds) for x in range(3)]
         torch_query_xyz = torch.stack(torch_query_xyz)
